+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @donkeylabs/cli
 
+## 2.0.3
+
+### Patch Changes
+
+- @donkeylabs/server@2.0.3
+
 ## 2.0.2
 
 ### Patch Changes